// src/app/core/services/model/tblPermiso.ts
/**
 * Mi API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TblRolesPermiso } from './tblRolesPermiso';


export interface TblPermiso {
    id?: number;
    nombre?: string | null;
    tblRolesPermisos?: Array<TblRolesPermiso> | null;
    isChecked: boolean
}